Breaking modules down further and offering an Android module w/ appropriate testing.
