+++ ucum.xmlpull/src/main/resources/META-INF/services/org.fhir.ucum.serializing.DefinitionParserProvider
@@ -1,0 +1,1 @@
+org.fhir.ucum.serializing.xmlpull.DefinitionParserProvider
