--- ucum.xpp3/src/main/resources/META-INF/services/org.fhir.ucum.serializing.DefinitionParserProvider
@@ -1,1 +1,0 @@
-org.fhir.ucum.serializing.xpp3.DefinitionParserProvider
